requirements.txt
sanic==0.7.0
sanic-script==0.3
sanic-mongodb-extension==0.1.1
sanic-amqp-extension==0.1.0

aioamqp==0.10.0
umongo==0.15.0
motor==1.2.1
marshmallow==2.15.0
sage-utils==0.5.4

pytest==3.4.0
pytest-cov==2.5.1
pytest-sanic==0.1.9

flake8==3.5.0
